requirements.txt
pandas==2.0.3
numpy==1.25.0
matplotlib==3.7.3
opencv-python-headless==4.8.0.76
pillow==9.5.0
pytesseract==0.3.10
imutils==0.5.4
tensorflow==2.14.0
tensorflow-hub==0.16.0